+++ Kosdas/HtmlExtension.cs
@@ -4,9 +4,9 @@ namespace Kosdas
 {
     public static class HtmlExtension
     {
-        public static decimal? ParseCell(this HtmlNode td)
+        public static double? ParseCell(this HtmlNode td)
         {
-            var parsed = decimal.TryParse(td.InnerText, out decimal value);
+            var parsed = double.TryParse(td.InnerText, out double value);
             if (parsed)
                 return value;
             return null;
